refactor(cockpit): move CockpitRouter + rail logic to cockpit_rail (#404)
Lift ``CockpitRouter``, ``CockpitItem`` and the pure rail-registry
walker helpers (``_selected_project_key``, ``_hidden_rail_items``,
``_collapsed_rail_sections``, ``_visibility_passes``,
``_rows_for_registration``) out of the monolithic ``cockpit.py`` and
into ``cockpit_rail.py`` where they sit next to the text renderer
(``PollyCockpitRail``) that already consumed them.

The split keeps the rail concerns — which row is selected, how state
is persisted, how a selection mounts in the right pane — in one
module that no longer has to cycle through ``pollypm.cockpit``.

Importers updated:

* src/pollypm/cli.py
* src/pollypm/cockpit_ui.py
* src/pollypm/cockpit_workers.py
* tests/test_cockpit.py
* tests/test_core_rail_items.py
* tests/test_rail_badges_visibility.py
* tests/test_rail_cli.py
* tests/integration/test_persona_integration.py

``CockpitRouter._rail_registry`` now resolves
``_register_worker_roster_rail_item`` / ``_register_metrics_rail_item``
via a deferred import so we don't reintroduce a module cycle while
those gather helpers still live in ``cockpit.py`` (they move in #405
along with the inbox + worker-roster extraction).

Acceptance: ``grep -n 'CockpitRouter\|CockpitItem' src/pollypm/cockpit.py``
returns zero hits. The 238-test cockpit/rail/palette/inbox/worker/metrics/
activity/keyboard suite passes locally.
