+++ src/pollypm/cockpit.py
@@ -1,19 +1,13 @@
 from __future__ import annotations
 
 import json
-import shlex
-from collections.abc import Callable
-from dataclasses import dataclass, replace
+from dataclasses import dataclass
 from pathlib import Path
 
-from pollypm.atomic_io import atomic_write_json
 from pollypm.config import load_config
-from pollypm.providers import get_provider
 from pollypm.projects import ensure_project_scaffold
-from pollypm.runtimes import get_runtime
 from pollypm.service_api import PollyPMService
 from pollypm.task_backends import get_task_backend
-from pollypm.session_services import create_tmux_client
 from pollypm.worktrees import list_worktrees
 
 # ---------------------------------------------------------------------------
@@ -321,1207 +315,6 @@ def _render_inbox_panel(config) -> str:
                 svc.close()
             except Exception:  # noqa: BLE001
                 pass
-
-
-
-@dataclass(slots=True)
-class CockpitItem:
-    key: str
-    label: str
-    state: str
-    selectable: bool = True
-
-
-def _selected_project_key(selected: object) -> str | None:
-    """Extract the project key from the ``selected`` cockpit state."""
-    if not isinstance(selected, str) or not selected.startswith("project:"):
-        return None
-    parts = selected.split(":", 2)
-    if len(parts) < 2:
-        return None
-    return parts[1] or None
-
-
-def _hidden_rail_items(config: object) -> frozenset[str]:
-    """Return user-configured hidden rail item keys (``section.label``)."""
-    rail_cfg = getattr(config, "rail", None)
-    hidden = getattr(rail_cfg, "hidden_items", None) if rail_cfg is not None else None
-    if not hidden:
-        return frozenset()
-    return frozenset(str(item) for item in hidden)
-
-
-def _collapsed_rail_sections(config: object) -> frozenset[str]:
-    """Return user-configured collapsed section names."""
-    rail_cfg = getattr(config, "rail", None)
-    collapsed = (
-        getattr(rail_cfg, "collapsed_sections", None) if rail_cfg is not None else None
-    )
-    if not collapsed:
-        return frozenset()
-    return frozenset(str(item) for item in collapsed)
-
-
-def _visibility_passes(reg, ctx) -> bool:
-    """Evaluate the registration's visibility predicate.
-
-    * ``"always"`` — always visible.
-    * ``"has_feature"`` — visible only if ``ctx.extras["features"]``
-      (a set/frozenset of capability names) includes
-      ``reg.feature_name`` (or ``reg.item_key`` as fallback).
-    * ``Callable`` — invoked; exceptions treat as hidden-and-logged.
-    """
-    import logging
-
-    visibility = reg.visibility
-    if visibility == "always":
-        return True
-    if visibility == "has_feature":
-        features = ctx.extras.get("features") or frozenset()
-        if not isinstance(features, (set, frozenset)):
-            try:
-                features = frozenset(features)
-            except TypeError:
-                return False
-        target = reg.feature_name or reg.item_key
-        return target in features
-    if callable(visibility):
-        try:
-            return bool(visibility(ctx))
-        except Exception:  # noqa: BLE001
-            logging.getLogger(__name__).exception(
-                "Rail item %s visibility predicate raised — hiding item",
-                reg.item_key,
-            )
-            return False
-    return True
-
-
-def _rows_for_registration(reg, ctx) -> list:
-    """Produce the list of :class:`RailRow` a registration renders to.
-
-    Default: one row using the (possibly dynamic) label, icon, and
-    state. When ``rows_provider`` is set we defer to the plugin —
-    handy for sections like ``projects`` where one registration fans
-    out into N rows.
-    """
-    import logging
-    from pollypm.plugin_api.v1 import RailRow
-
-    logger = logging.getLogger(__name__)
-
-    if reg.rows_provider is not None:
-        try:
-            rows = reg.rows_provider(ctx)
-        except Exception:  # noqa: BLE001
-            logger.exception(
-                "Rail item %s rows_provider raised — skipping", reg.item_key,
-            )
-            return []
-        return [r for r in rows if isinstance(r, RailRow)]
-
-    label = reg.label
-    if reg.label_provider is not None:
-        try:
-            dynamic_label = reg.label_provider(ctx)
-        except Exception:  # noqa: BLE001
-            logger.exception(
-                "Rail item %s label_provider raised — falling back to static",
-                reg.item_key,
-            )
-            dynamic_label = None
-        if isinstance(dynamic_label, str) and dynamic_label:
-            label = dynamic_label
-
-    state = "idle"
-    if reg.state_provider is not None:
-        try:
-            dynamic_state = reg.state_provider(ctx)
-        except Exception:  # noqa: BLE001
-            logger.exception(
-                "Rail item %s state_provider raised — falling back to idle",
-                reg.item_key,
-            )
-            dynamic_state = None
-        if isinstance(dynamic_state, str) and dynamic_state:
-            state = dynamic_state
-
-    # Badge appended to label if provider returns a non-null value. The
-    # badge-rendering tick is cheap; provider exceptions fall back to no
-    # badge per er03 acceptance.
-    if reg.badge_provider is not None:
-        try:
-            badge = reg.badge_provider(ctx)
-        except Exception:  # noqa: BLE001
-            logger.exception(
-                "Rail item %s badge_provider raised — rendering without badge",
-                reg.item_key,
-            )
-            badge = None
-        if badge not in (None, 0, ""):
-            # Only append a badge when the label_provider hasn't already
-            # baked the count in (e.g. "Inbox (3)" from core_rail_items).
-            if f"({badge})" not in label:
-                label = f"{label} ({badge})"
-
-    key = reg.selection_key
-    return [RailRow(key=key, label=label, state=state, selectable=True)]
-
-
-class CockpitRouter:
-    _STATE_FILE = "cockpit_state.json"
-    _COCKPIT_WINDOW = "PollyPM"
-    _LEFT_PANE_WIDTH = 30  # default; actual value persisted in cockpit state.
-
-    def __init__(self, config_path: Path) -> None:
-        self.config_path = config_path
-        self.service = PollyPMService(config_path)
-        self.tmux = create_tmux_client()
-        self._supervisor = None
-        # Per-project activity cache keyed by project key.
-        # value: (db_mtime, git_mtime, is_active, has_working_task)
-        # Skips re-opening SQLite on every 0.8s cockpit tick when nothing changed.
-        self._project_activity_cache: dict[str, tuple[float, float, bool, bool]] = {}
-
-    def _load_supervisor(self, *, fresh: bool = False):
-        # Reload config if the file changed (picks up new projects, sessions, etc.)
-        if not fresh and self._supervisor is not None:
-            try:
-                config_mtime = self.config_path.stat().st_mtime
-                if not hasattr(self, "_config_mtime") or config_mtime != self._config_mtime:
-                    fresh = True
-                    self._config_mtime = config_mtime
-            except OSError:
-                pass
-        if fresh or self._supervisor is None:
-            if self._supervisor is not None:
-                self._supervisor.store.close()
-            self._supervisor = self.service.load_supervisor()
-            self._supervisor.ensure_layout()
-            try:
-                self._config_mtime = self.config_path.stat().st_mtime
-            except OSError:
-                pass
-            # Bump epoch so the cockpit TUI refreshes on next tick
-            try:
-                from pollypm.state_epoch import bump
-                bump()
-            except Exception:  # noqa: BLE001
-                pass
-        return self._supervisor
-
-    def _state_path(self) -> Path:
-        config = load_config(self.config_path)
-        config.project.base_dir.mkdir(parents=True, exist_ok=True)
-        return config.project.base_dir / self._STATE_FILE
-
-    def selected_key(self) -> str:
-        self._validate_state()
-        data = self._load_state()
-        value = data.get("selected")
-        return str(value) if isinstance(value, str) and value else "polly"
-
-    def set_selected_key(self, key: str) -> None:
-        self._validate_state()
-        data = self._load_state()
-        data["selected"] = key
-        self._write_state(data)
-
-    def _load_state(self) -> dict[str, object]:
-        path = self._state_path()
-        if not path.exists():
-            return {}
-        try:
-            payload = json.loads(path.read_text())
-        except json.JSONDecodeError:
-            return {}
-        return payload if isinstance(payload, dict) else {}
-
-    def _write_state(self, data: dict[str, object]) -> None:
-        atomic_write_json(self._state_path(), data)
-
-    def rail_width(self) -> int:
-        """Return the persisted rail width, falling back to the default."""
-        data = self._load_state()
-        value = data.get("rail_width")
-        if isinstance(value, int) and 20 <= value <= 120:
-            return value
-        return self._LEFT_PANE_WIDTH
-
-    def set_rail_width(self, width: int) -> None:
-        """Persist the rail width so subsequent launches and layout checks use it."""
-        if not isinstance(width, int) or width < 20 or width > 120:
-            return
-        data = self._load_state()
-        if data.get("rail_width") == width:
-            return
-        data["rail_width"] = width
-        self._write_state(data)
-
-    def _validate_state(self, *, panes: list | None = None, target: str | None = None) -> list:
-        """Clear stale entries from cockpit_state.json.
-
-        Checks that right_pane_id points to a real pane and that
-        mounted_session is actually alive. Prevents stale state from
-        blocking heartbeat recovery or causing wrong session mounts.
-
-        Returns the list of panes fetched (or the list passed in) so
-        callers can avoid re-issuing ``list_panes`` — see #175.
-        """
-        state = self._load_state()
-        dirty = False
-        if target is None:
-            config = load_config(self.config_path)
-            target = f"{config.project.tmux_session}:{self._COCKPIT_WINDOW}"
-        if panes is None:
-            panes = self._safe_list_panes(target)
-
-        right_pane_id = state.get("right_pane_id")
-        right_pane = None
-        if isinstance(right_pane_id, str) and right_pane_id:
-            right_pane = next((pane for pane in panes if pane.pane_id == right_pane_id), None)
-            if right_pane is None or getattr(right_pane, "pane_dead", False):
-                state.pop("right_pane_id", None)
-                state.pop("mounted_session", None)
-                dirty = True
-                right_pane = None
-
-        mounted = state.get("mounted_session")
-        if isinstance(mounted, str) and mounted:
-            release_lease = False
-            try:
-                supervisor = self._load_supervisor()
-                launches = supervisor.plan_launches()
-                launch = next((l for l in launches if l.session.name == mounted), None)
-                if launch is None or not self._mounted_session_matches_pane(launch, right_pane):
-                    state.pop("mounted_session", None)
-                    dirty = True
-                    release_lease = True
-            except Exception:  # noqa: BLE001
-                state.pop("mounted_session", None)
-                dirty = True
-                release_lease = True
-            if release_lease:
-                self._release_cockpit_lease(supervisor if "supervisor" in locals() else None, mounted)
-
-        if dirty:
-            self._write_state(state)
-        return panes
-
-    def _safe_list_panes(self, target: str) -> list:
-        try:
-            return self.tmux.list_panes(target)
-        except Exception:  # noqa: BLE001
-            return []
-
-    def _mounted_session_matches_pane(self, launch, pane) -> bool:
-        if pane is None or getattr(pane, "pane_dead", False):
-            return False
-        if not self._is_live_provider_pane(pane):
-            return False
-        # A live provider pane is running — trust the state rather than
-        # trying to match CWD.  CWD matching is unreliable because the
-        # agent may cd elsewhere during its turn.  If state says this
-        # session is mounted and the pane is alive, believe it.
-        return True
-
-    def _release_cockpit_lease(self, supervisor, session_name: str) -> None:
-        if supervisor is None:
-            try:
-                supervisor = self._load_supervisor()
-            except Exception:  # noqa: BLE001
-                return
-        try:
-            supervisor.release_lease(session_name, expected_owner="cockpit")
-        except Exception:  # noqa: BLE001
-            pass
-
-    def _ui_initialized_sessions(self) -> set[str]:
-        data = self._load_state()
-        value = data.get("ui_initialized_sessions")
-        if not isinstance(value, list):
-            return set()
-        return {item for item in value if isinstance(item, str) and item}
-
-    def _mark_ui_initialized(self, session_name: str) -> None:
-        data = self._load_state()
-        current = data.get("ui_initialized_sessions")
-        items = [item for item in current if isinstance(item, str) and item] if isinstance(current, list) else []
-        if session_name not in items:
-            items.append(session_name)
-        data["ui_initialized_sessions"] = items
-        self._write_state(data)
-
-    def build_items(self, *, spinner_index: int = 0) -> list[CockpitItem]:
-        """Build rail rows by walking the plugin-host rail registry.
-
-        Rows are gathered in section order (``top`` → ``projects`` →
-        ``workflows`` → ``tools`` → ``system``), then within each
-        section by ``(index, plugin_name)``. Items that declare a
-        ``rows_provider`` expand into N rows; others collapse into a
-        single :class:`CockpitItem` built from the static label plus
-        optional ``label_provider`` / ``state_provider`` callables.
-
-        Pre-er02 behaviour is preserved by the built-in
-        ``core_rail_items`` plugin, which registers every rail entry
-        that used to be hardcoded here.
-        """
-        from pollypm.plugin_api.v1 import RailContext, RAIL_SECTIONS
-
-        supervisor = self._load_supervisor()
-        config = supervisor.config
-        launches, windows, alerts, _leases, _errors = supervisor.status()
-
-        cockpit_state = self._load_state()
-        ctx = RailContext(
-            selected_project=_selected_project_key(cockpit_state.get("selected")),
-            cockpit_state=dict(cockpit_state),
-            extras={
-                "router": self,
-                "supervisor": supervisor,
-                "config": config,
-                "launches": launches,
-                "windows": windows,
-                "alerts": alerts,
-                "spinner_index": spinner_index,
-            },
-        )
-
-        registry = self._rail_registry()
-
-        # Hidden items + visibility predicates land in er03 / er04. The
-        # renderer here runs them every tick — badge providers likewise
-        # (a crash falls back to no-badge).
-        hidden_keys = _hidden_rail_items(config)
-        collapsed_sections = _collapsed_rail_sections(config)
-
-        grouped: dict[str, list[CockpitItem]] = {name: [] for name in RAIL_SECTIONS}
-
-        for reg in registry.items():
-            if reg.item_key in hidden_keys:
-                continue
-            if not _visibility_passes(reg, ctx):
-                continue
-            rows = _rows_for_registration(reg, ctx)
-            for row in rows:
-                item = CockpitItem(
-                    key=row.key,
-                    label=row.label,
-                    state=row.state,
-                    selectable=row.selectable,
-                )
-                grouped.setdefault(reg.section, []).append(item)
-
-        items: list[CockpitItem] = []
-        for section in RAIL_SECTIONS:
-            rows = grouped.get(section) or []
-            if not rows:
-                continue
-            if section in collapsed_sections:
-                # Collapsed sections render as a disabled header row so
-                # the user can still see the section exists. Expansion
-                # is a runtime concept tracked via set_selected_key.
-                items.append(
-                    CockpitItem(
-                        key=f"_section:{section}",
-                        label=f"{section.upper()} (collapsed)",
-                        state="separator",
-                        selectable=False,
-                    )
-                )
-                continue
-            items.extend(rows)
-
-        return items
-
-    def _rail_registry(self):
-        """Return the plugin-host rail registry for the active root dir."""
-        from pollypm.plugin_host import extension_host_for_root
-
-        config = load_config(self.config_path)
-        host = extension_host_for_root(str(config.project.root_dir.resolve()))
-        # Initialize plugins so the rail registry is populated. Safe to
-        # call repeatedly — it tracks which plugins have been init'd.
-        try:
-            host.initialize_plugins(config=config)
-        except Exception:  # noqa: BLE001
-            # Plugin init failures surface via degraded_plugins; don't
-            # block the rail rendering.
-            pass
-        registry = host.rail_registry()
-        # Worker roster top-rail entry. Registered here (not in
-        # ``core_rail_items``) so the cockpit router + roster Textual
-        # app land in one feature drop — the rail row, the route handler
-        # (``route_selected("workers")``), and the panel renderer all
-        # live alongside each other. The registration is gated on
-        # ``core_rail_items`` being active so disabling the core plugin
-        # still yields an empty rail (see
-        # ``test_removing_core_rail_items_yields_empty_rail``).
-        try:
-            core_enabled = "core_rail_items" in host.plugins()
-        except Exception:  # noqa: BLE001
-            core_enabled = True
-        if core_enabled:
-            _register_worker_roster_rail_item(registry, self)
-            _register_metrics_rail_item(registry, self)
-        return registry
-
-    def _project_session_map(self, launches) -> dict[str, str]:
-        project_session_map: dict[str, str] = {}
-        for launch in launches:
-            if launch.session.role in {"operator-pm", "heartbeat-supervisor", "triage", "reviewer"}:
-                continue
-            project_session_map.setdefault(launch.session.project, launch.session.name)
-        return project_session_map
-
-    # Alert types that are informational / auto-managed — don't show red triangle
-    _SILENT_ALERT_TYPES = frozenset({
-        "suspected_loop",      # auto-clears when snapshot changes
-        "stabilize_failed",    # stale after successful recovery
-        "needs_followup",      # informational, handled by heartbeat
-    })
-
-    def _session_state(self, session_name: str, launches, windows, alerts, spinner_index: int) -> str:
-        actionable = [
-            a for a in alerts
-            if a.session_name == session_name and a.alert_type not in self._SILENT_ALERT_TYPES
-        ]
-        if actionable:
-            # Include a short reason so the user knows what's wrong
-            top = actionable[0]
-            short_reason = top.alert_type.replace("_", " ")
-            return f"! {short_reason}"
-        launch = next((item for item in launches if item.session.name == session_name), None)
-        if launch is None:
-            return "idle"
-        window_map = {window.name: window for window in windows}
-        window = window_map.get(launch.window_name)
-        # If the session is mounted in the cockpit, its storage window is gone.
-        # Check the cockpit right pane instead.
-        if window is None:
-            state = self._load_state()
-            if state.get("mounted_session") == session_name:
-                window = self._mounted_window_proxy(launch, windows)
-        if window is None:
-            return "idle"
-        if window.pane_dead:
-            return "dead"
-        spinners = ["\u25dc", "\u25dd", "\u25de", "\u25df"]
-        if launch.session.role in ("worker", "operator-pm", "reviewer"):
-            working = self._is_pane_working(window, launch.session.provider)
-            if working:
-                return spinners[spinner_index % 4] + " working"
-            if launch.session.role == "worker":
-                return "\u25cf live"
-            return "ready"
-        if launch.session.role == "heartbeat-supervisor":
-            return "watch"
-        if launch.session.role == "triage":
-            return "triage"
-        return "live"
-
-    def _mounted_window_proxy(self, launch, windows):
-        """Return a window-like object for a session mounted in the cockpit pane."""
-        cockpit_windows = [w for w in windows if w.name == self._COCKPIT_WINDOW]
-        if not cockpit_windows:
-            return None
-        # The cockpit window has multiple panes; the right pane is the mounted session.
-        try:
-            supervisor = self._load_supervisor()
-            target = f"{supervisor.config.project.tmux_session}:{self._COCKPIT_WINDOW}"
-            panes = self.tmux.list_panes(target)
-            if len(panes) < 2:
-                return None
-            right_pane = max(panes, key=self._pane_left)
-            # Return the cockpit window but with the right pane's info
-            cockpit_win = cockpit_windows[0]
-            from dataclasses import replace as dc_replace
-            return dc_replace(
-                cockpit_win,
-                pane_id=right_pane.pane_id,
-                pane_current_command=right_pane.pane_current_command,
-                pane_dead=right_pane.pane_dead,
-            )
-        except Exception:  # noqa: BLE001
-            return None
-
-    def _is_pane_working(self, window, provider) -> bool:
-        """Check if a session pane has an active turn (agent is working, not idle at prompt)."""
-        try:
-            pane_text = self.tmux.capture_pane(window.pane_id, lines=15)
-        except Exception:  # noqa: BLE001
-            return False
-        stripped = pane_text.rstrip()
-        if not stripped:
-            return False
-        tail = stripped[-200:]
-        lowered = tail.lower()
-        # Universal working indicator — both Claude and Codex show this during active turns
-        if "esc to interrupt" in lowered:
-            return True
-        # Universal idle indicators — if any are present, the session is idle regardless of provider
-        idle_markers = (
-            "bypass permissions", "new task?", "/clear to save", "shift+tab to cycle",  # Claude
-            "press enter to confirm", "% left",  # Codex
-        )
-        if any(marker in lowered for marker in idle_markers):
-            return False
-        # Provider-specific prompt detection
-        provider_value = provider.value if hasattr(provider, "value") else str(provider)
-        if provider_value == "claude":
-            return "\u276f" not in tail
-        if provider_value == "codex":
-            # Codex idle: › prompt
-            if "\u203a" in tail:
-                return False
-            return bool(stripped)
-        return False
-
-    def _cleanup_duplicate_windows(self, storage_session: str) -> None:
-        """Kill duplicate windows in the storage closet, keeping only the first of each name."""
-        try:
-            windows = self.tmux.list_windows(storage_session)
-        except Exception:  # noqa: BLE001
-            return
-        seen: dict[str, int] = {}  # name -> first index
-        for window in windows:
-            if window.name in seen:
-                try:
-                    self.tmux.kill_window(f"{storage_session}:{window.index}")
-                except Exception:  # noqa: BLE001
-                    pass
-            else:
-                seen[window.name] = window.index
-
-    def ensure_cockpit_layout(self) -> None:
-        config = load_config(self.config_path)
-        target = f"{config.project.tmux_session}:{self._COCKPIT_WINDOW}"
-        # Single list-panes baseline shared with ``_validate_state``. See
-        # #175: subsequent list-panes calls only run after a mutation that
-        # invalidates the cached view (split/kill); pure swaps preserve
-        # pane IDs so we update order locally.
-        panes = self._safe_list_panes(target)
-        self._validate_state(panes=panes, target=target)
-        # Clean up duplicate windows in the storage closet before layout setup
-        try:
-            supervisor = self._load_supervisor()
-            self._cleanup_duplicate_windows(supervisor.storage_closet_session_name())
-        except Exception:  # noqa: BLE001
-            pass
-        state = self._load_state()
-        right_pane_id = state.get("right_pane_id")
-        right_pane_present = isinstance(right_pane_id, str) and any(pane.pane_id == right_pane_id for pane in panes)
-        if len(panes) == 1 and right_pane_present and panes[0].pane_id == right_pane_id:
-            # The rail (left) pane died, only the worker (right) pane survived.
-            # Park the worker back to storage and clear stale state so the
-            # split below creates a fresh right pane from the cockpit pane.
-            supervisor = self._load_supervisor()
-            mounted = state.get("mounted_session")
-            if isinstance(mounted, str) and mounted:
-                launch = next(
-                    (item for item in supervisor.plan_launches() if item.session.name == mounted),
-                    None,
-                )
-                storage_session = supervisor.storage_closet_session_name()
-                if launch is not None and self.tmux.has_session(storage_session):
-                    try:
-                        self.tmux.break_pane(panes[0].pane_id, storage_session, launch.window_name)
-                    except Exception:  # noqa: BLE001
-                        pass
-            state.pop("right_pane_id", None)
-            state.pop("mounted_session", None)
-            self._write_state(state)
-            try:
-                panes = self.tmux.list_panes(target)  # structural change (break_pane)
-            except Exception:  # noqa: BLE001
-                panes = []
-            right_pane_id = None
-            right_pane_present = False
-        if len(panes) < 2:
-            # Calculate right pane size so the rail starts at exactly rail_width
-            # columns — avoids the visible flash of a 50/50 split followed by resize.
-            window_width = panes[0].pane_width if panes else 200
-            right_size = max(window_width - self.rail_width() - 1, 40)
-            right_pane_id = self.tmux.split_window(
-                target,
-                self._right_pane_command("polly"),
-                horizontal=True,
-                detached=True,
-                size=right_size,
-            )
-            state["right_pane_id"] = right_pane_id
-            self._write_state(state)
-            panes = self.tmux.list_panes(target)  # split added a pane
-        elif len(panes) > 2:
-            for pane in panes:
-                if pane.pane_id == panes[0].pane_id:
-                    continue
-                try:
-                    self.tmux.kill_pane(pane.pane_id)
-                except Exception:  # noqa: BLE001
-                    pass
-            panes = self.tmux.list_panes(target)  # kill_pane removed panes
-        if len(panes) >= 2:
-            # ``_normalize_layout`` may swap pane positions but never changes
-            # pane IDs or count — so we can reason about the post-swap left/
-            # right locally without another ``list-panes`` round-trip.
-            self._normalize_layout(target, panes)
-            left_pane, right_pane = self._post_normalize_lr(panes)
-            state["right_pane_id"] = right_pane.pane_id
-            self._write_state(state)
-            self._try_resize_rail(left_pane.pane_id)
-
-    def _post_normalize_lr(self, panes):
-        """Return ``(left, right)`` for the panes after ``_normalize_layout``.
-
-        ``_normalize_layout`` guarantees the ``uv`` (rail) pane ends up on
-        the left. When neither pane is ``uv`` (edge case), fall back to the
-        pre-swap ``pane_left`` ordering — the same answer the old code would
-        have computed via a second ``list-panes`` + ``min/max(pane_left)``.
-        """
-        if len(panes) != 2:
-            left = min(panes, key=self._pane_left)
-            right = max(panes, key=self._pane_left)
-            return left, right
-        a, b = panes
-        a_cmd = getattr(a, "pane_current_command", "")
-        b_cmd = getattr(b, "pane_current_command", "")
-        if a_cmd == "uv":
-            return a, b
-        if b_cmd == "uv":
-            return b, a
-        left = min(panes, key=self._pane_left)
-        right = max(panes, key=self._pane_left)
-        return left, right
-
-    def _pane_left(self, pane) -> int:
-        return int(getattr(pane, "pane_left", 0))
-
-    def _try_resize_rail(self, pane_id: str) -> None:
-        """Best-effort resize of the rail pane. Never raises."""
-        try:
-            self.tmux.resize_pane_width(pane_id, self.rail_width())
-        except Exception:  # noqa: BLE001
-            pass
-
-    def _right_pane_size(self, window_target: str) -> int | None:
-        """Calculate the exact right-pane size so the rail starts at rail_width columns."""
-        try:
-            panes = self.tmux.list_panes(window_target)
-            if panes:
-                window_width = max(p.pane_width for p in panes)
-                return max(window_width - self.rail_width() - 1, 40)
-        except Exception:  # noqa: BLE001
-            pass
-        return None
-
-    def _normalize_layout(self, target: str, panes) -> None:
-        if len(panes) != 2:
-            return
-        left_pane = min(panes, key=self._pane_left)
-        right_pane = max(panes, key=self._pane_left)
-        left_command = getattr(left_pane, "pane_current_command", "")
-        right_command = getattr(right_pane, "pane_current_command", "")
-        if left_command == "uv":
-            return
-        if right_command == "uv":
-            self.tmux.swap_pane(right_pane.pane_id, left_pane.pane_id)
-
-    def _right_pane_id(self, target: str) -> str | None:
-        panes = self.tmux.list_panes(target)
-        if len(panes) < 2:
-            return None
-        return max(panes, key=self._pane_left).pane_id
-
-    def _left_pane_id(self, target: str) -> str | None:
-        panes = self.tmux.list_panes(target)
-        if not panes:
-            return None
-        return min(panes, key=self._pane_left).pane_id
-
-    def route_selected(self, key: str) -> None:
-        supervisor = self._load_supervisor()
-        window_target = f"{supervisor.config.project.tmux_session}:{self._COCKPIT_WINDOW}"
-        self.ensure_cockpit_layout()
-        right_pane = self._right_pane_id(window_target)
-        if right_pane is None:
-            raise RuntimeError("Cockpit right pane is not available.")
-
-        self.set_selected_key(key)
-        if key == "polly":
-            # Launch operator session if not running
-            launches = supervisor.plan_launches()
-            storage_session = supervisor.storage_closet_session_name()
-            op_launch = next((l for l in launches if l.session.name == "operator"), None)
-            if op_launch is not None:
-                storage_windows = {w.name for w in self.tmux.list_windows(storage_session)}
-                if op_launch.window_name not in storage_windows:
-                    try:
-                        supervisor.launch_session("operator")
-                    except Exception:
-                        pass
-            try:
-                self._show_live_session(supervisor, "operator", window_target)
-            except Exception:
-                self._show_static_view(supervisor, window_target, "polly")
-            return
-        if key == "russell":
-            # Launch reviewer session if not running
-            launches = supervisor.plan_launches()
-            storage_session = supervisor.storage_closet_session_name()
-            rev_launch = next((l for l in launches if l.session.name == "reviewer"), None)
-            if rev_launch is not None:
-                storage_windows = {w.name for w in self.tmux.list_windows(storage_session)}
-                if rev_launch.window_name not in storage_windows:
-                    try:
-                        supervisor.launch_session("reviewer")
-                    except Exception:
-                        pass
-            try:
-                self._show_live_session(supervisor, "reviewer", window_target)
-            except Exception:
-                self._show_static_view(supervisor, window_target, "polly")
-            return
-        if key == "inbox":
-            self._show_static_view(supervisor, window_target, "inbox")
-            return
-        if key == "workers":
-            self._show_static_view(supervisor, window_target, "workers")
-            return
-        if key == "metrics":
-            self._show_static_view(supervisor, window_target, "metrics")
-            return
-        if key == "settings":
-            self._show_static_view(supervisor, window_target, "settings")
-            return
-        if key == "activity" or key.startswith("activity:"):
-            # Live activity feed — served through the standard static-view
-            # plumbing. ``activity:<project_key>`` preloads the per-project
-            # filter so the dashboard's ``l`` keybinding lands the user
-            # exactly where they were looking.
-            project_key = None
-            if key.startswith("activity:"):
-                _, _, project_key = key.partition(":")
-                project_key = project_key or None
-            self._show_static_view(
-                supervisor, window_target, "activity", project_key,
-            )
-            return
-        if key.startswith("project:"):
-            parts = key.split(":")
-            project_key = parts[1]
-            sub_view = parts[2] if len(parts) > 2 else None
-            if sub_view is None or sub_view == "dashboard":
-                # Clicking project parent or Dashboard — show dashboard,
-                # and set selection to the dashboard sub-item
-                self.set_selected_key(f"project:{project_key}:dashboard")
-                self._show_static_view(supervisor, window_target, "project", project_key)
-                return
-            if sub_view in ("settings", "issues"):
-                self._show_static_view(supervisor, window_target, sub_view, project_key)
-                return
-            if sub_view == "task" and len(parts) > 3:
-                # Per-task worker session — mount it from storage closet
-                task_num = parts[3]
-                window_name = f"task-{project_key}-{task_num}"
-                storage = supervisor.storage_closet_session_name()
-                try:
-                    storage_windows = self.tmux.list_windows(storage)
-                    target_win = next((w for w in storage_windows if w.name == window_name), None)
-                    if target_win is not None:
-                        self._park_mounted_session(supervisor, window_target)
-                        self._cleanup_extra_panes(window_target)
-                        left_pane = self._left_pane_id(window_target)
-                        right_pane_id = self._right_pane_id(window_target)
-                        if right_pane_id is not None:
-                            self.tmux.kill_pane(right_pane_id)
-                        source = f"{storage}:{target_win.index}.0"
-                        self.tmux.join_pane(source, left_pane, horizontal=True)
-                        panes = self.tmux.list_panes(window_target)
-                        left_p = min(panes, key=self._pane_left)
-                        self._try_resize_rail(left_p.pane_id)
-                        right_p = max(panes, key=self._pane_left)
-                        self.tmux.set_pane_history_limit(right_p.pane_id, 200)
-                        state = self._load_state()
-                        state["mounted_session"] = window_name
-                        state["right_pane_id"] = right_p.pane_id
-                        self._write_state(state)
-                        return
-                except Exception:  # noqa: BLE001
-                    pass
-                # Fallback to dashboard
-                self.set_selected_key(f"project:{project_key}:dashboard")
-                self._show_static_view(supervisor, window_target, "project", project_key)
-                return
-            if sub_view == "session":
-                # PM Chat — mount live session, spawning if needed
-                launches = supervisor.plan_launches()
-                session_name = self._project_session_map(launches).get(project_key)
-                if session_name is not None:
-                    if not self._session_available_for_mount(supervisor, session_name, window_target):
-                        # Session exists in config but not running — launch it
-                        try:
-                            supervisor.launch_session(session_name)
-                        except Exception:
-                            pass
-                    try:
-                        self._show_live_session(supervisor, session_name, window_target)
-                    except Exception:
-                        # Mount failed (e.g. duplicate windows) — fall back to dashboard
-                        self.set_selected_key(f"project:{project_key}:dashboard")
-                        self._show_static_view(supervisor, window_target, "project", project_key)
-                else:
-                    # No session configured — try to create one
-                    try:
-                        self.create_worker_and_route(project_key)
-                    except Exception:
-                        # Creation failed (e.g. no git commits) — fall back to dashboard
-                        self.set_selected_key(f"project:{project_key}:dashboard")
-                        self._show_static_view(supervisor, window_target, "project", project_key)
-                return
-            # Unknown sub_view — fall back to dashboard
-            self.set_selected_key(f"project:{project_key}:dashboard")
-            self._show_static_view(supervisor, window_target, "project", project_key)
-            return
-        raise RuntimeError(f"Unknown cockpit item: {key}")
-
-    def focus_right_pane(self) -> None:
-        config = load_config(self.config_path)
-        window_target = f"{config.project.tmux_session}:{self._COCKPIT_WINDOW}"
-        self.ensure_cockpit_layout()
-        right_pane = self._right_pane_id(window_target)
-        if right_pane is not None:
-            self.tmux.select_pane(right_pane)
-
-    def create_worker_and_route(
-        self,
-        project_key: str,
-        on_status: Callable[[str], None] | None = None,
-    ) -> None:
-        supervisor = self._load_supervisor()
-        launches = supervisor.plan_launches()
-        session_name = self._project_session_map(launches).get(project_key)
-        storage_session = supervisor.storage_closet_session_name()
-        storage_windows = {w.name for w in self.tmux.list_windows(storage_session)}
-
-        target: str | None = None
-        if session_name is not None:
-            launch = next(l for l in launches if l.session.name == session_name)
-            if launch.window_name not in storage_windows:
-                _launch, target = supervisor.create_session_window(session_name, on_status=on_status)
-        else:
-            prompt = self.service.suggest_worker_prompt(project_key=project_key)
-            self.service.create_and_launch_worker(
-                project_key=project_key, prompt=prompt, on_status=on_status, skip_stabilize=True,
-            )
-            # Re-read launches to pick up the newly created session
-            supervisor = self._load_supervisor(fresh=True)
-            launches = supervisor.plan_launches()
-            session_name = self._project_session_map(launches).get(project_key)
-            if session_name is not None:
-                launch = next(l for l in launches if l.session.name == session_name)
-                tmux_session = supervisor.tmux_session_for_launch(launch)
-                window_map = supervisor.window_map()
-                if launch.window_name in window_map:
-                    target_key = f"{tmux_session}:{launch.window_name}"
-                    # Window exists but hasn't been stabilized yet
-                    target = target_key
-
-        # Route immediately so the user sees the session booting live
-        self.route_selected(f"project:{project_key}")
-
-        # Stabilize in the background (dismisses prompts, waits for ready)
-        if target is not None and session_name is not None:
-            launch = next(l for l in supervisor.plan_launches() if l.session.name == session_name)
-            supervisor.stabilize_launch(launch, target, on_status=on_status)
-
-    def _show_live_session(self, supervisor, session_name: str, window_target: str) -> None:
-        mounted_session = self._mounted_session_name(supervisor, window_target)
-        launch = next(item for item in supervisor.plan_launches() if item.session.name == session_name)
-        if isinstance(mounted_session, str) and mounted_session == session_name:
-            right_pane_id = self._right_pane_id(window_target)
-            if right_pane_id is not None:
-                panes = self.tmux.list_panes(window_target)
-                right_pane = max(panes, key=self._pane_left)
-                if self._is_live_provider_pane(right_pane):
-                    return
-        self._park_mounted_session(supervisor, window_target)
-        self._cleanup_extra_panes(window_target)
-        left_pane_id = self._left_pane_id(window_target)
-        if left_pane_id is None:
-            raise RuntimeError("Cockpit left pane is not available.")
-        right_pane_id = self._right_pane_id(window_target)
-        storage_session = supervisor.storage_closet_session_name()
-        storage_windows = {window.name for window in self.tmux.list_windows(storage_session)}
-        if launch.window_name not in storage_windows:
-            # Control sessions (operator, reviewer) should be respawned
-            # when missing — the user clicking on Polly expects to talk to
-            # Polly, not see a placeholder.
-            if launch.session.role in {"operator-pm", "reviewer"}:
-                try:
-                    supervisor.launch_session(session_name)
-                    storage_windows = {w.name for w in self.tmux.list_windows(storage_session)}
-                    if launch.window_name in storage_windows:
-                        if right_pane_id is not None:
-                            self.tmux.kill_pane(right_pane_id)
-                        source = f"{storage_session}:{launch.window_name}.0"
-                        self.tmux.join_pane(source, left_pane_id, horizontal=True)
-                        panes = self.tmux.list_panes(window_target)
-                        left_pane = min(panes, key=self._pane_left)
-                        self._try_resize_rail(left_pane.pane_id)
-                        right_pane = max(panes, key=self._pane_left)
-                        self.tmux.set_pane_history_limit(right_pane.pane_id, 200)
-                        state = self._load_state()
-                        state["mounted_session"] = session_name
-                        state["right_pane_id"] = right_pane.pane_id
-                        self._write_state(state)
-                        return
-                except Exception:  # noqa: BLE001
-                    pass  # Fall through to static view if relaunch fails
-            # Non-control sessions or failed relaunch — show static view
-            fallback_kind = "polly" if session_name == "operator" else "project"
-            fallback_target = launch.session.project if fallback_kind == "project" else None
-            if right_pane_id is None:
-                right_size = self._right_pane_size(window_target)
-                right_pane_id = self.tmux.split_window(
-                    left_pane_id,
-                    self._right_pane_command(fallback_kind, fallback_target),
-                    horizontal=True,
-                    detached=True,
-                    size=right_size,
-                )
-            else:
-                self.tmux.respawn_pane(right_pane_id, self._right_pane_command(fallback_kind, fallback_target))
-            state = self._load_state()
-            state.pop("mounted_session", None)
-            state["right_pane_id"] = self._right_pane_id(window_target)
-            self._write_state(state)
-            return
-        if right_pane_id is not None:
-            self.tmux.kill_pane(right_pane_id)
-        # Use window index to avoid ambiguity with duplicate window names
-        storage_windows = self.tmux.list_windows(storage_session)
-        target_window = next(
-            (w for w in storage_windows if w.name == launch.window_name),
-            None,
-        )
-        if target_window is None:
-            self._show_static_view(supervisor, window_target, "polly" if session_name == "operator" else "project")
-            return
-        source = f"{storage_session}:{target_window.index}.0"
-        self.tmux.join_pane(source, left_pane_id, horizontal=True)
-        panes = self.tmux.list_panes(window_target)
-        left_pane = min(panes, key=self._pane_left)
-        self._try_resize_rail(left_pane.pane_id)
-        right_pane = max(panes, key=self._pane_left)
-        self.tmux.set_pane_history_limit(right_pane.pane_id, 200)
-        state = self._load_state()
-        state["mounted_session"] = session_name
-        state["right_pane_id"] = right_pane.pane_id
-        self._write_state(state)
-        # Auto-claim a cockpit lease so the heartbeat won't send
-        # nudges while a human is viewing/typing in this session.
-        try:
-            supervisor.claim_lease(session_name, "cockpit", "mounted in cockpit")
-        except Exception:  # noqa: BLE001
-            pass  # Lease may conflict — best effort
-
-    def _should_boot_visible(self, launch) -> bool:
-        if launch.session.name in self._ui_initialized_sessions():
-            return False
-        return launch.session.provider.value in {"claude", "codex"}
-
-    def _launch_visible_session(self, supervisor, launch, window_target: str, left_pane_id: str, right_pane_id: str | None):
-        storage_session = supervisor.storage_closet_session_name()
-        for window in self.tmux.list_windows(storage_session):
-            if window.name == launch.window_name:
-                self.tmux.kill_window(f"{storage_session}:{window.index}")
-                break
-        visible_launch = launch
-        if launch.session.provider.value == "codex" and launch.initial_input:
-            provider = get_provider(launch.session.provider, root_dir=supervisor.config.project.root_dir)
-            runtime = get_runtime(launch.account.runtime, root_dir=supervisor.config.project.root_dir)
-            visible_command = provider.build_launch_command(launch.session, launch.account)
-            visible_launch = replace(
-                launch,
-                command=runtime.wrap_command(visible_command, launch.account, supervisor.config.project),
-                initial_input=visible_command.initial_input,
-                resume_marker=visible_command.resume_marker,
-                fresh_launch_marker=visible_command.fresh_launch_marker,
-            )
-        if right_pane_id is not None:
-            self.tmux.kill_pane(right_pane_id)
-        right_size = self._right_pane_size(window_target)
-        right_pane_id = self.tmux.split_window(
-            left_pane_id,
-            visible_launch.command,
-            horizontal=True,
-            detached=False,
-            size=right_size,
-        )
-        self.tmux.set_pane_history_limit(right_pane_id, 200)
-        self.tmux.pipe_pane(right_pane_id, visible_launch.log_path)
-        supervisor.stabilize_launch(visible_launch, right_pane_id)
-        return max(self.tmux.list_panes(window_target), key=self._pane_left)
-
-    def _park_mounted_session(self, supervisor, window_target: str) -> None:
-        state = self._load_state()
-        mounted_session = self._mounted_session_name(supervisor, window_target)
-        if not isinstance(mounted_session, str) or not mounted_session:
-            return
-        right_pane_id = self._right_pane_id(window_target)
-        if right_pane_id is None:
-            state.pop("mounted_session", None)
-            self._write_state(state)
-            self._release_cockpit_lease(supervisor, mounted_session)
-            return
-        right_pane = max(self.tmux.list_panes(window_target), key=self._pane_left)
-        if not self._is_live_provider_pane(right_pane):
-            state.pop("mounted_session", None)
-            self._write_state(state)
-            self._release_cockpit_lease(supervisor, mounted_session)
-            return
-        launch = next(item for item in supervisor.plan_launches() if item.session.name == mounted_session)
-        storage_session = supervisor.storage_closet_session_name()
-        before = {(window.index, window.name) for window in self.tmux.list_windows(storage_session)}
-        self.tmux.break_pane(right_pane_id, storage_session, launch.window_name)
-        after = self.tmux.list_windows(storage_session)
-        created = [window for window in after if (window.index, window.name) not in before]
-        if created:
-            self.tmux.rename_window(f"{storage_session}:{created[-1].index}", launch.window_name)
-        else:
-            for window in after:
-                if window.name == self._COCKPIT_WINDOW:
-                    self.tmux.rename_window(f"{storage_session}:{window.index}", launch.window_name)
-                    break
-        state.pop("mounted_session", None)
-        self._write_state(state)
-        self._release_cockpit_lease(supervisor, mounted_session)
-
-    # Roles that should NEVER be auto-detected as mounted via CWD fallback.
-    # These are background roles — if the user is looking at a pane, it's
-    # not the heartbeat.  Guessing wrong here causes cascading mis-parks.
-    _NEVER_MOUNT_ROLES = frozenset({"heartbeat-supervisor", "triage"})
-
-    # When CWD is ambiguous (multiple sessions share the same cwd), prefer
-    # the session the user is most likely interacting with.
-    _MOUNT_PRIORITY = {"operator-pm": 0, "reviewer": 1, "worker": 2}
-
-    def _mounted_session_name(self, supervisor, window_target: str) -> str | None:
-        state = self._load_state()
-        mounted_session = state.get("mounted_session")
-        if isinstance(mounted_session, str) and mounted_session:
-            return mounted_session
-        if not hasattr(self.tmux, "list_panes"):
-            return None
-        right_pane_id = self._right_pane_id(window_target)
-        if right_pane_id is None:
-            return None
-        panes = self.tmux.list_panes(window_target)
-        if len(panes) < 2:
-            return None
-        right_pane = max(panes, key=self._pane_left)
-        if not self._is_live_provider_pane(right_pane):
-            return None
-        # CWD fallback: find the best matching session, but NEVER guess
-        # heartbeat or triage — those are background roles that should
-        # never be mounted in the cockpit.  When multiple sessions share
-        # a CWD, prefer operator > reviewer > worker.
-        pane_path = str(Path(right_pane.pane_current_path).resolve())
-        best_match: tuple[int, str] | None = None  # (priority, session_name)
-        for launch in supervisor.plan_launches():
-            if launch.session.role in self._NEVER_MOUNT_ROLES:
-                continue
-            session_cwd = str(Path(launch.session.cwd).resolve())
-            if pane_path == session_cwd:
-                priority = self._MOUNT_PRIORITY.get(launch.session.role, 5)
-                if best_match is None or priority < best_match[0]:
-                    best_match = (priority, launch.session.name)
-        if best_match is not None:
-            state["mounted_session"] = best_match[1]
-            state["right_pane_id"] = right_pane.pane_id
-            self._write_state(state)
-            return best_match[1]
-        return None
-
-    def _is_live_provider_pane(self, pane) -> bool:
-        cmd = getattr(pane, "pane_current_command", "")
-        # Claude Code may report the version string (e.g. "2.1.98") as the
-        # current command instead of "claude" or "node".
-        if cmd in {"node", "claude", "codex"}:
-            return True
-        # Treat any version-like string (digits and dots) as a live Claude pane.
-        if cmd and all(c.isdigit() or c == "." for c in cmd):
-            return True
-        return False
-
-    def _session_available_for_mount(self, supervisor, session_name: str, window_target: str) -> bool:
-        """Return True only if the session is already running (mounted or in storage)."""
-        mounted = self._mounted_session_name(supervisor, window_target)
-        if mounted == session_name:
-            return True
-        launch = next((item for item in supervisor.plan_launches() if item.session.name == session_name), None)
-        if launch is None:
-            return False
-        storage_session = supervisor.storage_closet_session_name()
-        storage_windows = {window.name for window in self.tmux.list_windows(storage_session)}
-        return launch.window_name in storage_windows
-
-    def _show_static_view(
-        self,
-        supervisor,
-        window_target: str,
-        kind: str,
-        project_key: str | None = None,
-    ) -> None:
-        self._park_mounted_session(supervisor, window_target)
-        self._cleanup_extra_panes(window_target)
-        left_pane_id = self._left_pane_id(window_target)
-        if left_pane_id is None:
-            raise RuntimeError("Cockpit left pane is not available.")
-        right_pane_id = self._right_pane_id(window_target)
-        if right_pane_id is None:
-            right_pane_id = self.tmux.split_window(
-                left_pane_id,
-                self._right_pane_command(kind, project_key),
-                horizontal=True,
-                detached=True,
-                size=self._right_pane_size(window_target),
-            )
-        else:
-            self.tmux.respawn_pane(right_pane_id, self._right_pane_command(kind, project_key))
-        state = self._load_state()
-        state.pop("mounted_session", None)
-        state["right_pane_id"] = self._right_pane_id(window_target)
-        self._write_state(state)
-
-    def _cleanup_extra_panes(self, window_target: str) -> None:
-        """Kill any extra panes beyond the expected 2 (rail + right)."""
-        try:
-            panes = self.tmux.list_panes(window_target)
-        except Exception:  # noqa: BLE001
-            return
-        if len(panes) <= 2:
-            return
-        left_pane = min(panes, key=self._pane_left)
-        # Keep the leftmost (rail) and rightmost (content) panes, kill the rest
-        right_pane = max(panes, key=self._pane_left)
-        for pane in panes:
-            if pane.pane_id not in {left_pane.pane_id, right_pane.pane_id}:
-                try:
-                    self.tmux.kill_pane(pane.pane_id)
-                except Exception:  # noqa: BLE001
-                    pass
-
-    def _right_pane_command(self, kind: str, project_key: str | None = None) -> str:
-        root = shlex.quote(str(self.config_path.parent.resolve()))
-        import shutil
-        pm_cmd = "pm" if shutil.which("pm") else "uv run pm"
-        args = [pm_cmd, "cockpit-pane", shlex.quote(kind)]
-        if project_key is not None:
-            args.append(shlex.quote(project_key))
-        joined = " ".join(args)
-        return f"sh -lc 'cd {root} && {joined}'"
