+++ src/pollypm/cockpit_workers.py
@@ -23,7 +23,7 @@ from textual.binding import Binding
 from textual.containers import Vertical
 from textual.widgets import DataTable, Static
 
-from pollypm.cockpit import CockpitRouter
+from pollypm.cockpit_rail import CockpitRouter
 from pollypm.cockpit_alerts import _setup_alert_notifier
 from pollypm.cockpit_palette import _open_command_palette, _open_keyboard_help
 from pollypm.config import load_config
